refactor: components, contexts, hooks, types 폴더로 분리
